Use scoped text locator when making text assertions
