+++ src/Api/Concerns/MakesElementAssertions.php
@@ -6,6 +6,8 @@ namespace Pest\Browser\Api\Concerns;
 
 use Illuminate\Support\Str;
 use Pest\Browser\Api\Webpage;
+use Pest\Browser\Playwright\Locator;
+use Pest\Browser\Support\WithinContext;
 use PHPUnit\Framework\ExpectationFailedException;
 
 /**
@@ -46,9 +48,7 @@ trait MakesElementAssertions
     {
         $text = (string) $text;
 
-        $locator = $this->page->unstrict(
-            fn () => $this->page->getByText($text),
-        );
+        $locator = $this->getScopedTextLocator($text);
 
         foreach ($locator->all() as $element) {
             if ($element->isVisible()) {
@@ -70,9 +70,7 @@ trait MakesElementAssertions
     {
         $text = (string) $text;
 
-        $locator = $this->page->unstrict(
-            fn () => $this->page->getByText($text),
-        );
+        $locator = $this->getScopedTextLocator($text);
 
         foreach ($locator->all() as $element) {
             if ($element->isVisible()) {
@@ -566,5 +564,18 @@ trait MakesElementAssertions
         $text = (string) $text;
 
         return $this->assertSee($text);
+    }
+
+    /** Gets a text locator respecting the current scope when set. */
+    private function getScopedTextLocator(string $text): Locator
+    {
+        $scope = WithinContext::getScope();
+
+        if ($scope !== null) {
+            $scopedLocator = $this->page->locator($scope);
+            return $this->page->unstrict(fn () => $scopedLocator->getByText($text));
+        }
+
+        return $this->page->unstrict(fn () => $this->page->getByText($text));
     }
 }
